[azure] Add auto fomatting for metrics (#220)
The returned metrics for an container instance are now auto fomatted.
This means, if the metrics is returned with the unit "Bytes" we try to
format the values in a unit which fits more (KiB, MiB, GiB, etc.).

This commit also fixes a bug, where the tooltip for an chart wasn't
displayed correctly, when the chart with the Azure metrics was rendered
in the drawer. This was caused by the "overflowX" property in the Tabs
component. Without this property the tooltip is rendered correctly.