[harbor] Add support for multi-arch images (#224)
This commits adds better support for multi-arch images, which are saved
in Harbor. Until now, we only were showing the parent artifact for
multi-arch image, without the build history and vulnerabilities. Now we
are using the reference array from this artifact to get more information
for all of the child artifacts. In this case one artifact represents one
architecture/os for the image.

To get the vulnerabilities for these artifacts we had to add the
"X-Accept-Vulnerabilities" header to the requests. If this header isn't
present the vulnerabilities are missing for multi-arch images.