[core] Change options handling
This commit changes the options handling again. In #181 we removed the
time parameter and since then we always showed the absolute time in the
options button. It turned out that this was very unintuitive for users,
so that we re-add the "time" property and show the relative time (e.g.
when a user selects "Last 15 Minutes").

We replaced the Options component with a Toolbar component, so that we
only display one button to trigger the search. Before this we displayed
a refresh button and a search button, which confused users. They didn't
exactly know when to use which button. Now with only one button
displayed we use a similar approach like Grafana, Kibana, etc. that the
time is always refreshed when the time is relative and the user clicks
on the search button.

We also use the current url only to get the initial options for a
component. When the user changes the options, we adjust the url and
directly adjust the options state, instead of using the adjust url to
set the new options.