[klogs] Add option to sort fields (#226)
It is now possible to sort the list of selected fields in the klogs
plugin. For that a list of action buttons is displayed when the user
hovers over a field. These actions can be used to move a field one
position up or down or to delete it from the list of selected fields.