[azure] Add support for Virtual Machine Scale Sets (#228)
The Azure plugin supports Virtual Machine Scale Sets now. This means
that it is possible to view Virtual Machine Scale Sets, there metrics,
the virtual machines and there metrics within kobs. If the AKS cluster
is using VMSS you can also navigate from your cluster to the
corresponding Virtual Machine Scale Sets.

The panel for "virtualmachinescalesets" supports the following options:

- "list": To list all Virtual Machine Scale Sets for a list of resource
  groups.
- "details": To show the details of a Virtual Machine Scale Set.
- "virtualMachines": To show a list of all virtual machines which were
  created by a VMSS.
- "metrics": To show metrics for a Virtual Machine Scale Set or for a
  single virtual machine, when the name of the virtual machine is
  provided.

We also changed the way how metrics can be retrieved from the Azure API
and embedded into a panel. Therefore it is now required to provide a
list of "metricNames" and an "aggregationType" to display metrics.