[klogs] Highlight expanded row
An expanded row is now highlighted in the klogs plugin. Besides that we
are not using the "index" as "key" anymore, so that already expanded
rows are not expanded when switching to another page.

The same change was also applied for the Elasticsearch plugin, but
instead of using the "timestamp" of the document we are using the "_id"
of the document.