[resources] Highlight expanded row
The row of an container is now highlighted when expanded. So we now
follow this style in all components which are using expanded rows.

We also fixed a typo in several components where "setIsExpanded" was
typed as "setIsExpaned" and we removed the "rowIndex" which was not used
in all components which are using expanded rows. So that now all
implementations of expanded rows should be similar.