[resources] Use TableComposable instead of Table
We are now using the TableComposable component instead of the Table
component to display Kubernetes resources, so that we are finally
removed the Table component from all places were it was used.

We also adjusted the style of hoverable / selectable tables across
several plugins, so that the selected row (the row which is shown in the
details) is now correctly highlighted.