[core] Switch to zap for logging
We are now using "go.uber.org/zap" for logging instead of
"github.com/sirupsen/logrus".

It is now also possible to enrich each log line via a context. This way
we can add the "userID" and "requestID" to most of the log lines and we
can add additional fields to the logging context via the
"ContextWithValue" function. For that we should always use our custom
"github.com/kobsio/kobs/pkg/log" package instead of directly using zap.