[flux] Fix sync action
This commit fixes the sync action in the Flux plugin for Kustomizations
and Helm Releases. After #235 we forgot to adjust the sync action, so
that the cluster, namespace and name of the Kustomization / Helm Release
was "undefined" in the action. Therefore the API call failed, because
the resource was not found and an error was shown to the user.

We also changed the behaviour of the "Search" button, so that the
resources are always reloaded, when the user clicks the button.