[resources] Fix reload of resources
This fixes that resources were not reloaded when a user clicks on the
search button, without changing the selected clusters, namespaces or
resources. This bug was caused, because we forgot to add the time
parameters, which are returned within the search function, to the url.
So that the update of the options was not triggered via our "useEffect"
hook.

The same problem was also present in the applications plugin, which is
also fixed with this commit.