Rename and adjust GitHub Actions
We renamed our GitHub Action, so that the "Test" action is now named
"Continuous Integration" and the "Docker" action is now named
"Continuous Delivery".

We also updated the "Continuous Integration" action, so that the flows
for Node.js and Go are run in parallel.

The "Continuous Integration" action is now also triggered, when an file
in the ".github/workflows" directory is changed.