[klogs] Fix out of time range documents (#245)
This commit fixes a bug in the klogs plugin, where it could happen that
the returned documents were not in the users selected time range. This
was caused by the buckets we determine before the actual query, were the
bucket timestamp could be not within the selected time range.

We also improved the test coverage for the klogs plugin and unify the
style of the used "table" tests.