[core] Improve filtering in select components
This improves the filtering in select components, so that also entries
are matched, which are containing a string but do not start with the
search string provided by the user.

This commit runs prittier in the fixed files and fixes some eslint
errors. I also changed the key property on most of the components to
improve the rendering.