[jaeger] Save expanded spans state
We are now saving the state of all spans, if they are expanded or not.
This ways we can also render an expanded span, when the user scrolled
the span out of the viewport and when he later scrolls back to the
expanded span.

Before this change, the span was not expanded anymore, when the user
scrolls through the list of spans.