[jaeger] Add new "publicAddress" option (#246)
* [jaeger] Add new "publicAddress" option

The new "publicAddress" option in the Jaeger plugin configuration, can
be used to define the address of a Jaeger UI instance, which can be
accessed by the users. If this option is set, we add a link to each
trace, which allows the users to open the trace in the Jaeger UI.

* [jaeger] Save expanded spans state

We are now saving the state of all spans, if they are expanded or not.
This ways we can also render an expanded span, when the user scrolled
the span out of the viewport and when he later scrolls back to the
expanded span.

Before this change, the span was not expanded anymore, when the user
scrolls through the list of spans.