[opsgenie] Adjust permissions and add more actions
This commit adopts the permission schema from the Azure plugin for the
Opsgenie plugin. This means that the permissions can now be defined on a
user / team level in the corresponding CRs. To grant a user / team
access to an action the following values can be used:
"acknowledgeAlert", "closeAlert", "snoozeAlert", "resolveIncident" and
"closeIncident".

We also added two new actions for incidents, so that it is now possible
to resolve and close incidents via kobs.

We also improved the logic after an actions was triggered. This means if
a user calls an action we refresh the corresponding list of alerts /
incidents after 3 seconds, so that closed alerts / incidents
automatically disappear in the lists or that the are updated when
another action is triggered.