[core] Refactor "cluster" and "clusters" package (#263)
This commit refactors the "cluster" and "clusters" package. We
introduced some new interfaces and now creating "clients" in both
packages which must implement the interfaces to interact with a
Kubernetes cluster.

This was required to increase the test coverage in the future, by
mocking the clients to interact with all / one Kubernetes cluser.

With this change with had to adjust multiple packages and plugins where
the clusters client was used. For all plugins which are not interacting
with a Kubernetes cluster we removed the clusters client as parameter.