[applications] Improve tags support (#265)
It is now possible to filter applications by tags, next to clusters and
namespaces. We also show the tags now on the applications page infront
of the teams and links list.

To implement these features we added a new cache layer, like the one for
teams and the topology graph. This layer is implemented by the new
"tags" package. It is used to cache all tags for a configurable amount
of time to reduce the load on the Kubernetes API server and shorter
response times. This is necessary because to get the list of tags we
have to look at all applications in all clusters.

NOTE: Currently the filtering by tags only works for the gallery view.
As part of the next task from #253 we should decide if we want to
display applications which doesn't match the query in the topology chart
or if we just want to render them in a different style (lower
"opacity" value).

With this commit we also increased the test coverage for the
applications package to 100%.

Relates #253.