[applications] Add dashboards to dependencies (#270)
It is now possible to add dashboards to a dependency of an application.
This can be done via the new "dashboards" property in the dependency
object. The dashboards can be referenced similar to how it is done in
the application.

The dashboards are only shown in the topology graph, when the user
clicks on the edge from the application to the application it depends
on.

The intended usage of this feature is it, to monitor the relation of two
applications, e.g. request metrics between the application and the
application it depends on.