[applications] Support for external applications (#271)
This commit adds support for external applications. An external
application can be used to monitor your applications which are not
running within a Kubernetes cluster like managed databases.

To find external applications within kobs, you still can filter for the
cluster / namespace where the Application CR was created, but we will
then not show the cluster / namespace in the UI, instead we are showing
the term "external".

In the topology graph, external applications are rendered outside of the
cluster / namespace node, this means that they not have the cluster /
namespace as parent.