[helm] Add Helm Plugin
The new Helm plugin can be used to manage Helm releases via kobs. It is
possible to view Helm releases across clusters and namespaces, including
their values, history and template files.

When the Helm plugin is used within a panel in a dashboard it is
possible to show a list of helm releases or the history of a single helm
release.

Closes #254.