[azure] Improve cost management integration (#275)
Improve the cost management integration in the Azure plugin. The pie
chart for the actual costs uses the same style as the other charts in
kobs. We are now using the chart theme, color scheme and tooltip from
the core package, so that the chart fits the style of the other charts
in kobs and the Azure plugin.

Instead of selecting the time from a dropdown in the cost management
page, it is now possible to select a time range via the Toolbar
component from the core package, like it is done in other plugins.

It is now possible to show the actual costs in a panel on a dashboard.
For this the type "costmanagement" with the sub type "actualcosts" can
be selected in the Azure plugin options. The user can also set a scope
(name of a resource group). If the user doesn't set this value it will
default to "All".