[resources] Add "verbs" property for permissions (#276)
To define the permissions for a user / team it is now required to also
add a list of allowed "verbs". The list of verbs can have the following
entries: "get", "patch", "post", "delete" and "*". The special entry "*"
can be used to allow all verbs for a user.

The verbs can be used to have a better control about what users can do
with kobs. For example with "get" a user may be able to list all
resources, but since he doesn't have the "patch", "post" or "delete"
verb, he can not modify running resources.

We also added two new special entries for the "resources" list. These
are "pods/logs" and "pods/exec" which can be used to control if a user
can get the logs of a Pod or if the user can get a terminal for a Pod.
The "pods/exec" term can also be used to control if a user can download
/ upload a file from / to a Pod together with the "get" / "post" verbs.