Support multiple versions for the documentation
The documentation of kobs supports multiple versions now. This means
that a new version of the documentation is published on each commit to
the main branch and for every new tag. The user can then switch between
these version on the documentation site.

For this feature we are using "mike" and the versioning feature of
"Material for MkDocs", see:
https://squidfunk.github.io/mkdocs-material/setup/setting-up-versioning/

Before this commit is merged to the main branch we have to delete
everything from the gh-pages branch except the CNAME file. Because mike
would place the CNAME file in a sub directory we have to create / keep
it manually in the gh-pages branch.