[dashboards] Refactor defaults and placeholders (#279)
This is a breaking change for the defaults and placeholders handling in
dashboards. We removed the "defaults" option for dashboards, so that
users must now set the cluster / namespace in all plugins. Before this
they were optional in all plugins which had a cluster(s) or namespace(s)
option. As a replacement for this we introduced two new special
variables. These are "__cluster" and "__namespace" which containing the
cluster and namespace of an Application or a Team were the dashboard is
used.

We also changed the behaviour of placeholders, so that they are now more
like variables. This means placeholders must now be used with the
following syntax "{% .<placeholder-name> %}". The old "{{
.<placeholder-name> }}" syntax doesn't work anymore.

Another change is that the defaults (cluster / namespace) in the CRDs
are now replaced on the server side and not in the frontend anymore.
This shouldn't have any effect for the end user.