[helm] Remove helm.sh/helm/v3 package (#281)
We decided to remove the "helm.sh/helm/v3" package in the Helm plugin.
Instead we are using a custom logic to get the Helm releases from the
Kubernetes API. These custom logic involves the following steps:

- List all secrets with "owner=helm"
- Loop through all returned secrets and decode the "release" key
- Decompress the decoded release data an parse them into the Release
  struct

We decided to use our own logic, because we had some problems with the
official Helm package and our cluster logic. This leads to a not so
satisfying experience during the development.

Besides that we are now displaying the "values.yaml" file in the
templates section in the frontend and we are using "isInline" for the
Alert component when it is rendered in a Card component.