[core] Update CRDs from "v1beta1" to "v1" (#284)
We updated all our CRDs from version "v1beta1" to "v1". This effects the
Custom Resource Definition for Applications, Dashboards, Teams and
Users.

This was done in preparation of the v1.0.0 release, which marks the
first stable version of kobs. Until the release we do not expect any
breaking changes in the CRDs, but if they are necessary, we will do
them. After the v1.0.0 release changes to the CRDs should always be
backwards compatible.