[core] Add debug endpoints
The api server exposed some "/api/debug" endpoints now. These endpoints
can be used to get the "http/pprof" information and to dump an http
request. The debug endpoints can be enabled via the "--api.debug" flag.
By default these endpoints are disabled.

We also fixed some typos in the documentation for the Azure and Helm
plugin. In the Azure we had a wrong link for the permissions
documentation and in the Helm plugin the configuration section was
missing.

Besides the changes from above, we also enabled the secure and http only
flag for the cookie, which is created by the auth middleware when
authentication is enabled.