[resources] Refactor forbidden resources config
We refactored the configuration for forbidden resources, to use the same
permission schema as it is used for the user and team permissions. This
means instead of just setting the resource name it is now possible to
restrict the access to resources based on the cluster name, namespace,
resource and verb.