Refactor permission handling for resources (#287)
* [resources] Refactor "pods/*" permissions

The special permissions "pods/log" and "pods/exec" can now only be used
together with the "*" as value for the verb. When the formaly valid
values "get" or "post" are used, a user can not access the logs or
terminal anymore.

* [resources] Refactor forbidden resources config

We refactored the configuration for forbidden resources, to use the same
permission schema as it is used for the user and team permissions. This
means instead of just setting the resource name it is now possible to
restrict the access to resources based on the cluster name, namespace,
resource and verb.

* [resources] Run go mod tidy