[resources] Add support to show custom columns
It is now possible to customize the columns which should be shown for a
resource via the new "columns" field in the resource panel options.

To show a set of custom columns for a resource, a user can set a lost of
columns. For each column the "title", "resource" and "jsonPath" must be
provided. The "title" is used as header for the column. The "resource"
field is used to apply the custom style only for the specifed resource
in a list of resources. The "jsonPath" field is used to select the value
for the column from the resource manifest file. The optional "type"
field can be used to apply a special formatter for the field value
(currently only "date" is supported as special formatter).