[klogs] Show links in dropdown
Instead of showing the links to the aggragations page and the
documentation via a simple text component on the logs page, we are now
using a actions menu which can be used to access the aggragations page
and the documentation.

We applied the same style for the aggregation page in the klogs plugin,
the Elasticsearch plugin and the Jaeger plugin.