[resources] Fix shown conditions in details view
The conditions where not shown correctly in the details view of a
resource. This happend because we filtered the conditions by the status
field when this was set to "True". We now do not filter the conditions
and instead strike through conditions which are not "True" to show them
all.

Besides that the node conditions were missing, which is now fixed and
the same fix was applied for the Flux plugin.