[azure] Allow custom value for tail and timestamps
It is now possible to set custom values for the tail and timestamps
paramter in the container instances panel. When a user wants to show the
logs of an container from a container group in a dashboard panel it is
now possible to set custom values for the "tail" and "timestamps"
paramter.

The "tail" paramter can be used to set the number of log lines which
should be shown (default "10000"). The "timestamps" paramter can be used
to show a timestamp infront of every log line (default "false").

We also fixed two bugs:

- When a user selected a container in logs tab, the select box wasn't
  closed. This is now fixed and after the user selects a container the
  options are not shown anymore.
- The second fix is in the mock client for the cost management
  integration, were the wrong version of the "costmanagement" package
  was used. With this fix we also added the missing tests for the cost
  management routes.