Improve shutdown of kobs (#294)
We are now waiting up to 60 seconds before all http servers of kobs are
stopped instead of the 5 seconds from before. This was done, because we
have some plugins which are having requests which could take a bit
longer, e.g. gettings logs via the klogs plugin.

We also added the "terminationGracePeriodSeconds" property to the
deployment for kobs in the Kustomize files and in the Helm chart. Since
we are waiting 60 seconds within kobs we set this value to 90 seconds.