[sql] Add columns option (#295)
It is now possible to format the returned column values for a SQL query.
For that a new "columns" property is available in the panel options,
which takes a map as value, where the key is the name of the column as
it is returned by the query and a columns object which allows users to
customize the title of a column and to format the column values.

We also fixed a bug, so that it is now possible to use the special time
variables within a SQL panel. For that we are now saving the index of
the selected query instead of the selected query themself.