[sql] Add support to visualize data
It is now possible to viasualize the data retrieved from an SQL instance
via line and area charts. For that a new type "chart" was introduced and
a "chart" property in the panel options. These options can be used by a
user to create line and area charts for his data.