[prometheus] Fix metrics with "Inf" values (#297)
When the used Prometheus client returns a metric with "Inf" values, we
were not able to able to parse it as JSON and show the results in the
UI. This is now fixed by checking if the value is "Inf". If this is the
case we just add the timestamp but not the value to our data structure
in the same way how we did it for "NaN" values.