Fix creation of jobs
Jobs are now created by using the complete Job template from a CronJob,
so that we are also using the defined metadata like labels and
annotations for a created Job.

Before this changed we just copied all the specs from the Job template
of a CronJob, but didn't included the metadata.