Hide scrollbars
We are now hidding the scrollbars via css, with loosing the
functionality for scrolling as a next step for a cleaner ui. This should
be used across all the kobs plugins, as long as it is clear that there
might be more "text" which can be viewed when we scroll in a container.
If it isn't visible in the first place that there might be more text we
should still display the scrollbars.