[core] Disable unnecessary scrollbars (#308)
* Disable unnecessary scrollbars

We displayed a lot of scrollbars (especially in the dashboards) also
when it was not necessary. By changing the overflow property from
"scroll" to "auto" most of the scrollbars should be gone, so that we
have a cleanier ui.

* Hide scrollbars

We are now hidding the scrollbars via css, with loosing the
functionality for scrolling as a next step for a cleaner ui. This should
be used across all the kobs plugins, as long as it is clear that there
might be more "text" which can be viewed when we scroll in a container.
If it isn't visible in the first place that there might be more text we
should still display the scrollbars.