[dashboards] Disable Intersection Observer API (#312)
We decided to disable the Intersection Observer API for the case that a
row in dashboard contains the "size" property set to "-1".

The "size" property in a dashboard can be used to not set an "unlimited"
size for a row, with the goal that the row can take as much space as
needed to avoid unnecessary scrolling within the row panels. Since we
can not precalculate the height for such rows, we decided to disable the
usage of the Intersection Observer API for dashboards where the "size"
property is set to "-1", to not have some buggy scroll behaviour.