Rework kobs architecture
We decided to rework the kobs architecture, by splitting it into two
components. A "satellite" which can run in each Kubernetes cluster which
should be manageable via kobs and a "hub" which provides access to the
"satellites" and which serves the UI.

This should allow us to run a single kobs instance in the future, where
we do not have to expose the services behind plugins to the public.