Fix plugin type handling in hub router
The handling of the plugin type wasn't working, so that a user received
an unauthorized error, when a plugin route was called with enabled
authentication.

We are now using a type param in the chi route to get the plugin type
instead of getting it from the raw url path.