[jaeger] Fix links and colors
The details links in the actions of the Jaeger plugin were broken,
because of a leading slash ("/").

When we assigned the color "red" to a service, the border in the spans
chart and view was a bit to dark, so that we are now using a different
color for red.