[app] Fix plugin filter (#352)
This commit fixes the filter on the plugin instances page. We didn't
used the plugin id as key for rendering the plugin gallery items, which
lead to some strange behaviour, so that plugins were still shown also
when the filter was changed. This should be fixed now.

It is now also possible to filter the list of plugins by the name of a
satellite, so that only plugins are shown that are configured at the
specific satellite instance.