[app] Change URLs for Application / Dashboard page
Since our IDs are containing a slash ("/") to seperate the different
parts (satellite, cluster, namespace and name) we had to encode the ID
to use it as parameter in the React Router. This is not working
properly, so that it can happen that the encoded slash character is
replaced by a real slash when a user reloads an URL. This leads to the
situation that React Router isn't aware of this url format and so
nothing is rendered.

We decided to adjust the URLs in React Router, so that we can still use
slashes in our IDs, but instead of just a "type" parameter, we now have
a parameter for every id part, which results in the following routes:

- /applications/satellite/:satellite/cluster/:cluster/namespace/:namespace/name/:name
- /dashboards/satellite/:satellite/cluster/:cluster/namespace/:namespace/name/:name