[app] Add profile page (#355)
It is now possible to display a list of dashboards on the users profile
page. The dashboards can be configured by every user via the "dashboards"
field in the User CR, so that each user can have a custom profile page.
If we coudln't found a CR for the email address of the user an
administrator can add a list of default dashboards via the hub
configuration file. If no configuration for the dashboards is provided
we will use a default dashboards list, which contains two dashboards:
One to display the team, where a user is part of and another one to
display the applications which are owned by these teams.

We also improved the applications panel, which now supports pagination,
to reduce the amount of applications which are loaded at once and to
keep the list of applications shorter.