[app] Adjust namespace handling
When a user selects multiple clusters and wanted to view the same
namespace in all clusters, he had to select the namespace for each
selected cluster.

This is now changed so that we do not have a cluster specific namespace
list anymore. Instead the list of namespaces is a unique list of
namespaces across all selected clusters.