[app] Fix namespace handling (#358)
When a user selected multiple clusters and multiple namespaces, there
was a bug so that we didn't show a complete list of resources.

This bug was caused, by the ways how we build the namespace ids in the
frontend: For that we used the map feature over the clusters and
namespaces which resulted in a two dimensonal array. We are now using
"flat()" on this array to get the correct list of namespace ids.

We also sorting the list of namespaces before it is shown to the user
now.