[app] Fix pagination (#359)
Our implementation for pagination doesn't worked correctly, because the
page number wasn't adjusted when the number of items per page was
changed (e.g. user is on the last page and selects a higher number for
the items per page).

This is now fixed, by setting the page to "1" when the number of items
per page is changed.