[app] Fix static file handling
Until now we only checked if a requested url contains a dot (".") to
invoke our static file handler. This could lead to a problem, so that a
page was not correctly served when it also contains a dot, because we
assumed that it must be a static file.

Now we are checking if the requested path contains "/static/" or one of
the listed file extensions to invoke our static file handler. If none of
these conditions match we serve our React app.