[app] Add tracing support via Jaeger or Zipkin
kobs now supports tracing via the OpenTelemetry Jaeger or Zipkin
exporter.

At this point not all functions are instrumented, but most of them are
creating spans now. We also propergate the trace information on all
outgoing http requests and try to get them from incoming requests, so
that kobs can be seamlessly integregated with other applications which
are support this.