[app] Add topology graph
This commit re-adds the topology graph for kobs. The topology graph is
generated by the provided dependencies in an Application and rendered
via cytoscape.

Note: We decided to remove the topology "type" feature, so that the
style of the nodes is always the same and can not be customized anymore.