[app] Add topology graph (#363)
* [app] Add topology graph

This commit re-adds the topology graph for kobs. The topology graph is
generated by the provided dependencies in an Application and rendered
via cytoscape.

Note: We decided to remove the topology "type" feature, so that the
style of the nodes is always the same and can not be customized anymore.

* [app] Add details view for topology graph

When a user selects a node (application) in the topology graph, we try
to get the corresponding application and display it in the detais drawer
of the topology page.

* [app] Add topology panel

It is now possible to display the topology graph of an application
within a panel on a dashboard.