[app] Fix usage of cytoscape in topology graph
In the production build the topology graph was not rendered correctly,
because the "layout.current.run()" function wasn't triggered. We now
ensure that this function is executed by adding a timeout of 100ms
before it should be run.

The same fix was also applied for the "istio" and the "kiali" plugin,
were we had the same issue.