[app] Rework Go plugin architecture (#366)
We used the "plugin" mode for building the Go plugins. The idea was to
make it easy for users to use their own plugins within kobs.
Unfortunately this didn't work well, so that we decided to build the
plugins within the kobs binary. For this a user can now register his
plugins within the "cmd/kobs/main.go" file.

We also adjusted the Docker image to use Alpine as base image again and
to disable CGO in the build.