[app] Improve filter for select components (#367)
We are now using "item.includes(value)" to check if an item in a select
box matches the user provided filter term. This should improve the user
experience for filtering item, because the user provided filter value
can also be in the middle of a string and not just in the beginning.

It is now also possible to submit a query in the Applications page by
pressing enter in the search input field.

Finally the user must now select at least one cluster in the Topology
page to render the topology graph. We decided to add this requirement to
avoid large queries against the backend when a user opens the page.