[flux] Fix panel implementation
The panel implementation for the Flux plugin was not correct, so that
the plugin could not be used with a dashboard. This is now fixed and the
corresponding is adjusted.

We also fixed a small bug in the usage of the custom columns feature of
the built-in "resources" plugin and adjusted the corresponding
documentation.