[app] Fix race condition in WaitGroups
We had a race condition in the resources handler, introduced by our
usage of "sync.WaitGroup". In the spawned Goroutines we appended some
data to a global slice, which caused the race condition. This is now
fixed by adding a mutex, so that we can safely append data to the slice.

The same problem also occured in the [rss] plugin. There we fixed the
problem in a similar way.