[app] Change Toolbar Handling (#371)
We decided to replace the Patternfly "Toolbar" and "ToolbarItem"
component with a custom "Toolbar" and "ToolbarItem" component, which is
based on the "Flex" and "FlexItem" components of Patternfly. This allows
us to have a better toolbar handling on smaller screen size, where we
can wrap the toolbar items to a new line or display them in a column
layout.

We also adjusted the Favicon, Apple and Android icons. Besides that we
also added the icons and configuration for windows tiles.