[app] Change drawer panel behaviour (#373)
This commit changes the behaviour of the drawer panel on pages where we
show dashboards (application, team and profile page), so that the drawer
does not overlap with the dashboards tab bar. Instead it is now shown
below the tab bar with the dashboards.

We also changed the behaviour of the search button in the Options
component, so that the onClick event is also triggered when the user
selected a custom time range.

Last but not least we are also encoding the parameters (service,
operation, tags) in the Jaeger plugin now.