[app] Add ServiceWorker and version parameter
We are now using a ServiceWorker for the React app served by kobs.
Besides that we are also adding a "version" parameter for the
"remoteEntry.js" files of the plugins, so that they are not used from
cache when a new version of kobs is deployed. The version is set during
build time and returned to the frontend via the "/api/plugins" call
which now contains the plugin instances and the version.