[app] Replace interface{} with any (#378)
We decided to replace "interface{}" with "any". This was done using the
following command: "gofmt -w -r 'interface{} -> any' ."

We did not replace "interface{}" with "any" for all generated code, this
means the created mocks still use "interface{}" and the created Go
clients for our Custom Resource Definitions are also unsing "interface{}".