[app] Adjust documentation (#379)
This is the first part of adjustments for the kobs v0.9.0. We adjusted
the landing page, the GitHub readme and contributing guide.

The contributing section in the docs will be adjusted, once the v0.9.0
tag was created and the kobsio/plugin-template and kobsio/app-template
release for the new version was created.