[app] Fix tooltip in charts (#383)
The tooltip in charts was not working. The bug was reproducable in the
Azure, Istio, Kiali, klogs, Prometheus and SQL plugin. To fix it we had
to remove the "@nivo/core" dependencie from the package.json file.

We also imrpoved the layout of the tooltip, to look better at different
label sizes.