[app] Customize Navigation Sidebar (#388)
It is now possible to customize the navigation sidebar, by adding custom
navigation groups and items. Each navigation item can contain a list of
child items or a reference to a dashboard. The dashboard reference must
contain the satellite, cluster, namespace and name of the dashboard. It
can also contain a list of placeholders, which are passed to the
dashboard.

Instead of the proposed solution in #384 we did not use the Flyout
component, because the component was not working properly on mobile
devices. We decided to use expandable navigation items.

Closes #384.