[app] Add Notifications
It is now possible to show notifications within kobs. Notifications can
be configured using the hub configuration file via plugins. When
notifications are configured a notification bell icon is added to the
header in the kobs ui. When a user clicks on the icon a notification
drawer is shown with all the retrieved notification items from a plugin.

In the this first iteration only the Opsgenie plugin is supported for
notifications to show open alerts and incidents.

For later is is planed to also allow resources and maybe also RSS feeds
to be displayed in the notification section.